Remove sdwdate-gui-qubes@.socket's explicit dependency on sockets.target
systemd.socket(5) documents:

  Socket units automatically gain a Before= dependency on sockets.target.

However, we had 'Requires/After' dependencies on sockets.target.
Marek noticed that sockets.target and sdwdate-gui-qubes@1000.socket
were ending up in a shutdown ordering cycle that was resulting in the
stop job for user@1000.service being deleted during shutdown. This is
a possible cause of the shutdown hanging bugs we are seeing in Qubes
OS's testing infrastructure.

Get rid of the dependency on sockets.target entirely, just rely on the
default 'Before' dependency.